Rename files using kebab-case
