+++ apps/idos-enclave/src/lib/mpc/secretsharing/Lagrange.ts
@@ -1,5 +1,5 @@
-import type { FiniteFieldElement } from "./FiniteFieldElement";
-import { Polynomial } from "./Polynomial";
+import type { FiniteFieldElement } from "./finite-field-element";
+import { Polynomial } from "./polynomial";
 
 /**
  * Try to interpolate a polynomial that passes through the supplied points.
